+++ packages/dialtone-vue2/components/rich_text_editor/rich_text_editor.vue
@@ -7,18 +7,29 @@
 
 <script>
 import { Editor, EditorContent } from '@tiptap/vue-2';
+import Blockquote from '@tiptap/extension-blockquote';
 import CodeBlock from '@tiptap/extension-code-block';
 import Document from '@tiptap/extension-document';
 import HardBreak from '@tiptap/extension-hard-break';
 import Paragraph from '@tiptap/extension-paragraph';
 import Placeholder from '@tiptap/extension-placeholder';
+import Bold from '@tiptap/extension-bold';
+import BulletList from '@tiptap/extension-bullet-list';
+import Italic from '@tiptap/extension-italic';
+import TipTapLink from '@tiptap/extension-link';
+import ListItem from '@tiptap/extension-list-item';
+import OrderedList from '@tiptap/extension-ordered-list';
+import Strike from '@tiptap/extension-strike';
+import Underline from '@tiptap/extension-underline';
 import Text from '@tiptap/extension-text';
+import TextAlign from '@tiptap/extension-text-align';
 import Emoji from './extensions/emoji';
 import Link from './extensions/link';
 import { MentionPlugin } from './extensions/mentions/mention';
 import {
   RICH_TEXT_EDITOR_OUTPUT_FORMATS,
   RICH_TEXT_EDITOR_AUTOFOCUS_TYPES,
+  RICH_TEXT_EDITOR_SUPPORTED_LINK_PROTOCOLS,
 } from './rich_text_editor_constants';
 
 import suggestion from './extensions/mentions/suggestion';
@@ -46,6 +57,14 @@ export default {
     editable: {
       type: Boolean,
       default: true,
+    },
+
+    /**
+     * Whether the input allows for line breaks to be introduced in the text.
+     */
+    allowLineBreaks: {
+      type: Boolean,
+      default: false,
     },
 
     /**
@@ -175,7 +194,18 @@ export default {
   computed: {
     extensions () {
       // These are the default extensions needed just for plain text.
-      const extensions = [CodeBlock, Document, Paragraph, Text];
+      const extensions = [
+        Blockquote,
+        Bold,
+        BulletList,
+        Document,
+        Italic,
+        ListItem,
+        Paragraph,
+        Strike,
+        Text,
+        Underline,
+      ];
       if (this.link) {
         extensions.push(this.getExtension(Link, this.link));
       }
@@ -187,21 +217,28 @@ export default {
 
       // make sure that this is defined before any other extensions
       // where Enter and Shift+Enter should have its own interaction. otherwise it will be ignored
-      extensions.push(
-        HardBreak.extend({
-          addKeyboardShortcuts () {
-            return {
-              Enter: () => true,
-              'Shift-Enter': () => this.editor.commands.first(({ commands }) => [
-                () => commands.newlineInCode(),
-                () => commands.createParagraphNear(),
-                () => commands.liftEmptyBlock(),
-                () => commands.splitBlock(),
-              ]),
-            };
-          },
-        }),
-      );
+      if (!this.allowLineBreaks) {
+        extensions.push(
+          HardBreak.extend({
+            addKeyboardShortcuts () {
+              return {
+                Enter: () => true,
+                'Shift-Enter': () => this.editor.commands.first(({ commands }) => [
+                  () => commands.newlineInCode(),
+                  () => commands.createParagraphNear(),
+                  () => commands.liftEmptyBlock(),
+                  () => commands.splitBlock(),
+                ]),
+              };
+            },
+          }),
+        );
+      }
+
+      extensions.push(TipTapLink.extend({ inclusive: false }).configure({
+        autolink: true,
+        protocols: RICH_TEXT_EDITOR_SUPPORTED_LINK_PROTOCOLS,
+      }));
 
       if (this.mentionSuggestion) {
         // Add both the suggestion plugin as well as means for user to add suggestion items to the plugin
@@ -212,6 +249,21 @@ export default {
       // Emoji has some interactions with Enter key
       // hence this should be done last otherwise the enter wont add a emoji.
       extensions.push(Emoji);
+
+      extensions.push(TextAlign.configure({
+        types: ['paragraph'],
+        defaultAlignment: 'left',
+      }));
+
+      extensions.push(CodeBlock.configure({
+        HTMLAttributes: {
+          class: 'dt-rich-text-editor--code-block',
+        },
+      }));
+
+      extensions.push(OrderedList.configure({
+        itemTypeName: 'listItem',
+      }));
 
       return extensions;
     },
@@ -361,5 +413,24 @@ export default {
     color: var(--dt-color-foreground-placeholder);
     pointer-events: none;
     height: 0;
+  }
+
+  .ProseMirror ul > li {
+    list-style-type: disc;
+  }
+
+  .ProseMirror ol > li {
+    list-style-type: decimal;
+  }
+
+  .ProseMirror blockquote {
+    padding-left: var(--dt-space-400);
+    border-left: var(--dt-size-border-300) solid var(--dt-color-foreground-muted-inverted);
+    margin-left: 0;
+  }
+
+  .dt-rich-text-editor--code-block {
+    background: var(--dt-color-surface-secondary);
+    padding: var(--dt-space-400);
   }
 </style>