+++ packages/dialtone-vue2/components/stack/stack_constants.js
@@ -8,7 +8,7 @@ export const DT_STACK_DIRECTION = {
 
 export const DT_STACK_RESPONSIVE_BREAKPOINTS = ['sm', 'md', 'lg', 'xl'];
 
-export const DT_STACK_GAP = ['0', '100', '200', '300', '400', '500', '600'];
+export const DT_STACK_GAP = ['0', '100', '200', '300', '400', '450', '500', '600'];
 
 export default {
   DT_STACK_DIRECTION,