+++ packages/dialtone-vue3/recipes/conversation_view/editor/index.js
@@ -1,0 +1,2 @@
+export { default as DtRecipeWysiwygEditor } from './editor.vue';
+export {} from './editor_constants.js';